requirements.txt
apache-airflow
apache-airflow-backport-providers-amazon
boto3